refactor: put types in separate file
