+++ src/pages/profile/profile-orders/index.tsx
@@ -1,46 +1,11 @@
 import { useEffect, useState } from 'react';
+import api from '@services/api';
 import ProfileOrder from '@components/profile-components/profile-order';
 import ProfileOrderMobile from '@components/profile-components/profile-order-mobile';
 import ReturnBackButton from '@components/profile-components/return-back-button';
 import { useProfile } from '@hooks/use-profile';
+import type { CommonOrder } from '../types';
 import styles from './profile-orders.module.scss';
-import api from '@services/api';
-// import type { Product } from '@services/generated-api/data-contracts';
-// import { OrderList } from '@services/generated-api/data-contracts.ts';
-
-type OrderStatusType =
-	| 'Ordered'
-	| 'In processing'
-	| 'Collecting'
-	| 'Gathered'
-	| 'In delivering'
-	| 'Delivered'
-	| 'Completed';
-
-type Product = {
-	amount: number;
-	final_price: number;
-	id: number;
-	measure_unit: string;
-	name: string;
-	quantity: string;
-	photo: string;
-	category: {
-		category_name: string;
-		category_slug: string;
-	};
-};
-
-type CommonOrder = {
-	id: number;
-	order_number?: string;
-	ordering_date?: string;
-	total_price?: string;
-	payment_method?: string;
-	delivery_method?: string;
-	status?: OrderStatusType;
-	products: Array<{ product: Product; quantity: string }> | Product[];
-};
 
 export default function ProfileOrders() {
 	const [isOpenDetails, setIsOpenDetails] = useState<number>();
